pyproject.toml
[build-system]
requires = [
    "setuptools >= 40.6.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "classic-examples-posts"
version = "0.0.1"
authors = [
    {name = "Artem Variasov", email = "varias070@gmail.com"},
]
description = "Provides utilities for creating components"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "SQLAlchemy==2.0.43",
    "pydantic==2.11.9",
    "waitress==3.0.2",
    "falcon==3.1.3",
    "alembic==1.16.5",
    "psycopg2==2.9.10",
    "classic-operations==0.3.4",
    "classic-components==1.3.2",
    "classic-falcon-integration==1.0.1",
    "pydantic_settings==2.10.1"
]

[project.optional-dependencies]
dev = [
    "build~=1.2.2.post1",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
]

[tool.setuptools.packages.find]
where = ["sources"]